pyproject.toml
[project]
name = "wvoid-radio"
version = "0.1.0"
description = "WVOID-FM: 24/7 AI Radio Station with The Liminal Operator"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27.0",
    "mutagen>=1.47.0",
    "ffmpeg-normalize>=1.28.0",
    "pyyaml>=6.0.1",
    "paramiko>=3.4.0",
    "watchdog>=4.0.0",
    "numpy<2.4",
    "resemble-perth>=1.0.1",
    "numba>=0.60.0",
    "llvmlite>=0.40.0",
    "librosa>=0.11.0",
    "s3tokenizer>=0.3.0",
    "transformers>=4.57.3",
    "diffusers>=0.36.0",
    "conformer>=0.3.2",
    "safetensors>=0.7.0",
    "spacy-pkuseg>=1.0.1",
    "pykakasi>=2.3.0",
    "gradio>=6.2.0",
    "pyloudnorm>=0.1.1",
    "omegaconf>=2.3.0",
    "torchcodec>=0.9.1",
]

[project.optional-dependencies]
tts = [
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["mac"]

[dependency-groups]
dev = [
    "ruff>=0.1.0",
]
